# Flush once for enqueue / workflow handler assignment