--- a/lib/galaxy/jobs/manager.py
+++ b/lib/galaxy/jobs/manager.py
@@ -49,7 +49,7 @@ class JobManager(object):
     def _message_callback(self, job):
         return JobHandlerMessage(task='setup', job_id=job.id)
 
-    def enqueue(self, job, tool=None):
+    def enqueue(self, job, tool=None, flush=True):
         """Queue a job for execution.
 
         Due to the nature of some handler assignment methods which are wholly DB-based, the enqueue method will flush
@@ -76,7 +76,7 @@ class JobManager(object):
         message_callback = partial(self._message_callback, job)
         try:
             return self.app.job_config.assign_handler(
-                job, configured=configured_handler, queue_callback=queue_callback, message_callback=message_callback)
+                job, configured=configured_handler, queue_callback=queue_callback, message_callback=message_callback, flush=flush)
         except HandlerAssignmentError as exc:
             raise ToolExecutionError(exc.args[0], job=exc.obj)