--- a/lib/galaxy/web_stack/handlers.py
+++ b/lib/galaxy/web_stack/handlers.py
@@ -280,7 +280,7 @@ class ConfiguresHandlers(object):
 
     # If these get to be any more complex we should probably modularize them, or at least move to a separate class
 
-    def _assign_handler_direct(self, obj, configured):
+    def _assign_handler_direct(self, obj, configured, flush=True):
         """Directly assign a handler if the object has been preconfigured to a known single static handler.
 
         :param obj:             Same as :method:`ConfiguresHandlers.assign_handler()`.
@@ -295,11 +295,12 @@ class ConfiguresHandlers(object):
                 handlers = None
             if handlers == (configured,):
                 obj.set_handler(configured)
-                _timed_flush_obj(obj)
+                if flush:
+                    _timed_flush_obj(obj)
                 return configured
         return False
 
-    def _assign_mem_self_handler(self, obj, method, configured, queue_callback=None, **kwargs):
+    def _assign_mem_self_handler(self, obj, method, configured, queue_callback=None, flush=True, **kwargs):
         """Assign object to this handler using this process's in-memory queue.
 
         This method ignores all handler configuration.
@@ -319,11 +320,12 @@ class ConfiguresHandlers(object):
             log.warning("(%s) Ignoring handler assignment to '%s' because configured handler assignment method"
                         " '' overrides per-tool handler assignment", obj.log_str(),
                         HANDLER_ASSIGNMENT_METHODS.MEM_SELF, configured)
-        _timed_flush_obj(obj)
+        if flush():
+            _timed_flush_obj(obj)
         queue_callback()
         return self.app.config.server_name
 
-    def _assign_db_self_handler(self, obj, method, configured, **kwargs):
+    def _assign_db_self_handler(self, obj, method, configured, flush=True, **kwargs):
         """Assign object to this process by setting its ``handler`` column in the database to this process.
 
         This only occurs if there is not an explicitly configured handler assignment for the object. Otherwise, it is
@@ -340,10 +342,11 @@ class ConfiguresHandlers(object):
                 obj, method, configured, **kwargs
             )
         obj.set_handler(self.app.config.server_name)
-        _timed_flush_obj(obj)
+        if flush:
+            _timed_flush_obj(obj)
         return self.app.config.server_name
 
-    def _assign_db_preassign_handler(self, obj, method, configured, index=None, **kwargs):
+    def _assign_db_preassign_handler(self, obj, method, configured, index=None, flush=True, **kwargs):
         """Assign object to a handler by setting its ``handler`` column in the database to a handler selected at random
         from the known handlers in the appropriate tag.
 
@@ -372,7 +375,8 @@ class ConfiguresHandlers(object):
             log.debug("(%s) Selected handler '%s' by random choice from handler tag '%s'", obj.log_str(),
                       handler_id, handler)
         obj.set_handler(handler_id)
-        _timed_flush_obj(obj)
+        if flush:
+            _timed_flush_obj(obj)
         return handler_id
 
     def _assign_db_tag(self, obj, method, configured, **kwargs):
@@ -438,8 +442,8 @@ class ConfiguresHandlers(object):
         # that's currently the best place for it. It's worth noting that this method is also part of the
         # WorkflowSchedulingManager, which acts like a combined JobConfiguration and JobManager. Combining those two
         # classes would probably be reasonable (and would remove the need for the queue callback).
-        if self._assign_handler_direct(obj, configured):
-            log.info("(%s) Skipped handler assignment logic due to explicit configuration to a single handler: %s",
+        if self._assign_handler_direct(obj, configured, flush=kwargs.get('flush', True)):
+            log.info("(%s) Skipped handler assignment logic due to explicit configuration` to a single handler: %s",
                      obj.log_str(), configured)
             return True
         for method in self.handler_assignment_methods: