@jmchilton 2020-07-22
Looks great to me - should we ping Nate and see if any of the more esoteric handler assignment types require this flush? I thought I had a stray thought worried about that last time I looked at this.