@mvdbeek 2020-07-22
Yep, @natefoo do you see anything that would obviously break if we defer the flush for handler assignment as in this PR ?