@natefoo 2020-08-05
Hilarious you all think I'd have any memory of the intricacies of these things 2 years later. 😆 

This is only for workflow handler assignment, right? Which you should always use db-preassign for. Although I forget what happens if you configure handlers using db-skip-locked but don't have configured workflow schedulers.